+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Support false positives with TypeScript function overloading (fixes [#105](https://github.com/ArnaudBarre/eslint-plugin-react-refresh/issues/105))
+- Support nested function calls for extraHOCs (fixes [#104](https://github.com/ArnaudBarre/eslint-plugin-react-refresh/issues/104))
 
 ## 0.5.0